# Conflicts:
#	randogui.py

+++ sslib/allpatch.py
@@ -34,6 +34,9 @@ class AllPatcher:
         self.bzs_patch=None
         self.event_patch=None
         self.event_text_patch=None
+        def dummy_progress_callback(action):
+            pass
+        self.progress_callback = dummy_progress_callback
         if not (self.actual_extract_path / 'DATA').exists():
             raise Exception('actual extract path should have a DATA subdir, make sure the directory structure is properly set up!')
     
@@ -98,55 +101,57 @@ class AllPatcher:
             match = STAGE_REGEX.match(stagepath.parts[-1])
             stage = match[1]
             layer = int(match[2])
-            # TODO: skip if nothing is modified at all
+            self.progress_callback(f'patching {stage} l{layer}')
             modified_stagepath = self.modified_extract_path/'DATA'/'files'/'Stage'/f'{stage}'/f'{stage}_stg_l{layer}.arc.LZ'
             modified = False
-            stagedata = nlzss11.decompress(stagepath.read_bytes())
-            stageu8 = U8File.parse_u8(BytesIO(stagedata))
             # remove some arcs if necessary
             remove_arcs = set(self.stage_oarc_delete.get((stage, layer), []))
             # add additional arcs if needed
             additional_arcs = set(self.stage_oarc_add.get((stage, layer), []))
-            # remove arcs that are already added on layer 0
-            if layer != 0:
-                additional_arcs = additional_arcs - set(self.stage_oarc_add.get((stage, 0), []))
-            remove_arcs = remove_arcs - additional_arcs
-            for arc in remove_arcs:
-                stageu8.delete_file(f'oarc/{arc}.arc')
-                modified = True
-            for arc in additional_arcs:
-                oarc_bytes = (self.oarc_cache_path / f'{arc}.arc').read_bytes()
-                stageu8.add_file_data(f'oarc/{arc}.arc', oarc_bytes)
-                modified = True
-            if layer == 0:
-                stagebzs = parseBzs(stageu8.get_file_data('dat/stage.bzs'))
-                # patch stage
-                if self.bzs_patch:
-                    newstagebzs = self.bzs_patch(stagebzs,stage,None)
-                    if newstagebzs is not None:
-                        stageu8.set_file_data('dat/stage.bzs', buildBzs(newstagebzs))
-                        modified = True
-                    # patch rooms
-                    room_path_matches = (ROOM_REGEX.match(x) for x in stageu8.get_all_paths())
-                    room_path_matches = (x for x in room_path_matches if not x is None)
-                    for room_path_match in room_path_matches:
-                        roomid = int(room_path_match.group('roomid'))
-                        roomdata = stageu8.get_file_data(room_path_match.group(0))
-                        roomarc = U8File.parse_u8(BytesIO(roomdata))
-                        roombzs = parseBzs(roomarc.get_file_data('dat/room.bzs'))
-                        roombzs = self.bzs_patch(roombzs, stage, roomid)
-                        if roombzs is not None:
-                            roomarc.set_file_data('dat/room.bzs', buildBzs(roombzs))
-                            stageu8.set_file_data(room_path_match.group(0), roomarc.to_buffer())
+            if remove_arcs or additional_arcs or layer == 0:
+                # only decompress and extract files, if needed
+                stagedata = nlzss11.decompress(stagepath.read_bytes())
+                stageu8 = U8File.parse_u8(BytesIO(stagedata))
+                # remove arcs that are already added on layer 0
+                if layer != 0:
+                    additional_arcs = additional_arcs - set(self.stage_oarc_add.get((stage, 0), []))
+                remove_arcs = remove_arcs - additional_arcs
+                for arc in remove_arcs:
+                    stageu8.delete_file(f'oarc/{arc}.arc')
+                    modified = True
+                for arc in additional_arcs:
+                    oarc_bytes = (self.oarc_cache_path / f'{arc}.arc').read_bytes()
+                    stageu8.add_file_data(f'oarc/{arc}.arc', oarc_bytes)
+                    modified = True
+                if layer == 0:
+                    stagebzs = parseBzs(stageu8.get_file_data('dat/stage.bzs'))
+                    # patch stage
+                    if self.bzs_patch:
+                        newstagebzs = self.bzs_patch(stagebzs,stage,None)
+                        if newstagebzs is not None:
+                            stageu8.set_file_data('dat/stage.bzs', buildBzs(newstagebzs))
                             modified = True
+                        # patch rooms
+                        room_path_matches = (ROOM_REGEX.match(x) for x in stageu8.get_all_paths())
+                        room_path_matches = (x for x in room_path_matches if not x is None)
+                        for room_path_match in room_path_matches:
+                            roomid = int(room_path_match.group('roomid'))
+                            roomdata = stageu8.get_file_data(room_path_match.group(0))
+                            roomarc = U8File.parse_u8(BytesIO(roomdata))
+                            roombzs = parseBzs(roomarc.get_file_data('dat/room.bzs'))
+                            roombzs = self.bzs_patch(roombzs, stage, roomid)
+                            if roombzs is not None:
+                                roomarc.set_file_data('dat/room.bzs', buildBzs(roombzs))
+                                stageu8.set_file_data(room_path_match.group(0), roomarc.to_buffer())
+                                modified = True
             # repack u8 and compress it if modified
             if modified:
                 stagedata = stageu8.to_buffer()
                 write_bytes_create_dirs(modified_stagepath, nlzss11.compress(stagedata))
-                print(f'patched {stage} l{layer}')
+                # print(f'patched {stage} l{layer}')
             elif self.copy_unmodified:
                 shutil.copy(stagepath, modified_stagepath)
-                print(f'copied {stage} l{layer}')
+                # print(f'copied {stage} l{layer}')
 
         # events and text
         eventrootpath = None
@@ -163,6 +168,7 @@ class AllPatcher:
         for eventpath in eventrootpath.glob('*.arc'):
             modified = False
             filename = eventpath.parts[-1]
+            self.progress_callback(f'patching {filename}')
             modified_eventpath = modified_eventrootpath / filename
             eventarc = U8File.parse_u8(BytesIO(eventpath.read_bytes()))
             for eventfilepath in eventarc.get_all_paths():
@@ -183,4 +189,4 @@ class AllPatcher:
                             modified = True
             if modified:
                 write_bytes_create_dirs(modified_eventpath, eventarc.to_buffer())
-                print(f'patched {filename}')
+                # print(f'patched {filename}')
